requirements.txt
pure-python-adb
cv2
numpy
